requirements.txt
Django==1.8.5
django-extensions==1.5.7
django-filter==0.11.0
djangorestframework==3.2.4
six==1.10.0
slugify==0.0.1
